backend/requirements.txt

fastapi==0.104.1
uvicorn==0.23.2
scikit-learn==1.3.2
pandas==2.1.1
numpy==1.26.0
python-multipart==0.0.6
pydantic==2.4.2
supabase==2.0.2
python-dotenv==1.0.0
